feat: aggregate 조회 API와 SQL schema baseline 정리

+++ src/main/java/net/vivans/dcim/module/query/api/dto/LastPointValueResponse.java
@@ -1,11 +1,12 @@
 package net.vivans.dcim.module.query.api.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record LastPointValueResponse(
         String pointName,
         String unit,
-        Double value,
+        BigDecimal value,
         Instant time
 ) {
 }
